Support the same new ticket URL for anonymous and logged-in users (#2122)
## What

Updates the help center so that `/help?new_ticket=true` works for both
anonymous and logged in users - for anonymous it opens the ticket form,
and when logged in it redirects to the tickets page with the form open.

### Authenticated


https://github.com/user-attachments/assets/11247296-f036-4b89-8e37-57f6744cb860

### Unauthenticated


https://github.com/user-attachments/assets/7946235e-c011-4733-8fe1-b30e7c112f59

## Why

Currently when we redirect people from other channels (e.g. GitHub
issues) to our support system, we can't give people a consistent URL to
open the modal. This will make it easier for us to tell people how to
open tickets.