Prevent sending duplicate subscription cancellation emails to creator (#2198)
### What

Skips sending a duplicate
`ContactingCreatorMailer.subscription_autocancelled()` email to the
creator if it was sent for the same customer in recent times (i.e.
between a week ago and now) in cases where it calls
`Subscription#unsubscribe_and_fail!` repeatedly for various reasons such
as if customer reactivated a cancelled/failed subscription but it was
failed/cancelled again.

Co-authored-by: Ershad Kunnakkadan <ershad92@gmail.com>